% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_junction.R
\name{cdf_junction}
\alias{cdf_junction}
\title{Cumulative distribution function for each junction}
\usage{
cdf_junction(dataset, sd.x, sd.t, times, t0, x0, type)
}
\arguments{
\item{dataset}{Data for incoming links over the entire junction}

\item{sd.x}{Vector of standard deviations used for calculating the distribution function}

\item{sd.t}{Vector of standard deviations used for calculating the density function over differences in time}

\item{times}{Number of observations}

\item{t0}{Time points recorded corresponding to relevant sampling rate}

\item{x0}{Vector of quantiles}

\item{type}{CDF calculated with functional only (type = 2), distributional only (type = 1) and both functional and distributional (type = 3)}
}
\description{
Cumulative distribution function for each junction
}
\examples{
cdf_junction()

}
