% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closedness.R
\name{closedness}
\alias{closedness}
\title{Compute the closedness}
\usage{
closedness(cluster, adjacency)
}
\arguments{
\item{cluster}{cluster}

\item{adjacency}{Adjacency matrix}
}
\description{
This function calculates the dissimilarity between CDFs for junctions
}
\examples{
closedness()
}
