% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_badness.R
\name{compute_badness}
\alias{compute_badness}
\title{Compute the dissimilarity function}
\usage{
compute_badness(i, j, cdf.junction)
}
\arguments{
\item{i}{Junction i}

\item{j}{Junction j}

\item{cdf.junction}{List of cumulative distribution functions for junctions}
}
\description{
This function calculates the dissimilarity between CDFs for junctions
}
\examples{
compute_badness()
}
