% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_link.R
\name{junction_link}
\alias{junction_link}
\title{Clustering algorithm}
\usage{
junction_link(adjacency, clustering, cdf.junction)
}
\arguments{
\item{adjacency}{Matrix to represent links from junction j to junction i}

\item{clustering}{Set of clusters}

\item{cdf.junction}{Matrix to represent CDF for each junction}
}
\description{
This function clusters the network using CDFs calculated for individual junctions aggregated over relevant incoming links.
}
\examples{
junction_link()
}
