% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_allowed.R
\name{merge_allowed}
\alias{merge_allowed}
\title{Check if the two junctions within a cluster are adjacent}
\usage{
merge_allowed(i, j, clustering, adjacency)
}
\arguments{
\item{i}{Junction i}

\item{j}{Junction j}

\item{clustering}{Set of clusters}

\item{adjacency}{Matrix to represent links from Junction j to Junction i}
}
\description{
This function allows only adjacenct junctions to merge to ensure spatially contiguous clusters
}
\examples{
merge_allowed()
}
