% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_penalty.R
\name{shape_penalty}
\alias{shape_penalty}
\title{Compute the shape penalty function}
\usage{
shape_penalty(i, j, clustering)
}
\arguments{
\item{i}{Junction i}

\item{j}{Junction j}

\item{clustering}{List of cumulative distribution functions}
}
\description{
This function calculates the dissimilarity between CDFS for junctions
}
\examples{
shape_penalty()
}
